fix: enter押しっぱなしを禁止
